requirements.txt
onnx~=1.9.0
tf2onnx
onnxruntime~=1.7.2
numpy~=1.20.3
tensorflow~=2.0.0
torch
gin-config
torchvision~=0.8.2
h5py<3.0.0

pycocotools~=2.0.2
pillow~=8.3.1
scikit-image~=0.18.1